requirements.txt
# JointTrack - Confluent Hackathon Requirements

# Core Dependencies (Existing)
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
flask>=2.3.0
flask-cors>=4.0.0
boto3>=1.28.0
moviepy>=1.0.3

# New Dependencies for Confluent + Google Cloud
confluent-kafka>=2.3.0
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.38.0
websockets>=12.0  # For real-time feedback to frontend

# Optional but recommended
requests>=2.31.0
